Rename monomial_*.{h,cc} to symbolic_monomial_*.{h,cc}
