+++ drake/common/symbolic_monomial.h
