+++ drake/common/symbolic_monomial_util.h
@@ -10,8 +10,8 @@
 
 #include "drake/common/drake_assert.h"
 #include "drake/common/hash.h"
-#include "drake/common/monomial.h"
 #include "drake/common/symbolic_expression.h"
+#include "drake/common/symbolic_monomial.h"
 #include "drake/common/symbolic_variable.h"
 #include "drake/common/symbolic_variables.h"
