+++ drake/common/test/symbolic_polynomial_test.cc
@@ -5,8 +5,8 @@
 
 #include <gtest/gtest.h>
 
-#include "drake/common/monomial.h"
-#include "drake/common/monomial_util.h"
+#include "drake/common/symbolic_monomial.h"
+#include "drake/common/symbolic_monomial_util.h"
 #include "drake/common/test/symbolic_test_util.h"
 
 namespace drake {
