+++ drake/solvers/BUILD
@@ -53,10 +53,10 @@ drake_cc_library(
         ":decision_variable",
         "//drake/common",
         "//drake/common:autodiff",
-        "//drake/common:monomial",
         "//drake/common:number_traits",
         "//drake/common:polynomial",
         "//drake/common:symbolic",
+        "//drake/common:symbolic_monomial",
         "//drake/common:symbolic_polynomial",
     ],
 )
@@ -70,9 +70,9 @@ drake_cc_library(
         ":constraint",
         ":symbolic_extraction",
         "//drake/common",
-        "//drake/common:monomial",
         "//drake/common:polynomial",
         "//drake/common:symbolic",
+        "//drake/common:symbolic_monomial",
     ],
 )
 
@@ -85,9 +85,9 @@ drake_cc_library(
         ":cost",
         ":symbolic_extraction",
         "//drake/common",
-        "//drake/common:monomial",
         "//drake/common:polynomial",
         "//drake/common:symbolic",
+        "//drake/common:symbolic_monomial",
         "//drake/common:symbolic_polynomial",
         "//drake/common:unused",
     ],
@@ -196,10 +196,10 @@ drake_cc_library(
         ":symbolic_extraction",
         "//drake/common",
         "//drake/common:autodiff",
-        "//drake/common:monomial",
         "//drake/common:number_traits",
         "//drake/common:polynomial",
         "//drake/common:symbolic",
+        "//drake/common:symbolic_monomial",
         "//drake/common:symbolic_polynomial",
     ],
 )
