+++ drake/solvers/symbolic_extraction.cc
@@ -8,8 +8,8 @@
 #include <utility>
 
 #include "drake/common/eigen_types.h"
-#include "drake/common/monomial.h"
 #include "drake/common/symbolic_expression.h"
+#include "drake/common/symbolic_monomial.h"
 #include "drake/math/matrix_util.h"
 
 namespace drake {
