+++ drake/solvers/test/mathematical_program_test.cc
@@ -19,10 +19,10 @@
 #include "drake/common/drake_assert.h"
 #include "drake/common/drake_copyable.h"
 #include "drake/common/eigen_matrix_compare.h"
-#include "drake/common/monomial.h"
 #include "drake/common/polynomial.h"
 #include "drake/common/symbolic_expression.h"
 #include "drake/common/symbolic_formula.h"
+#include "drake/common/symbolic_monomial.h"
 #include "drake/common/symbolic_polynomial.h"
 #include "drake/common/symbolic_variable.h"
 #include "drake/common/test/is_dynamic_castable.h"
