+++ drake/solvers/test/sos_constraint_test.cc
@@ -2,9 +2,9 @@
 
 #include <gtest/gtest.h>
 
-#include "drake/common/monomial_util.h"
 #include "drake/common/symbolic_expression.h"
 #include "drake/common/symbolic_formula.h"
+#include "drake/common/symbolic_monomial_util.h"
 #include "drake/common/symbolic_polynomial.h"
 #include "drake/common/symbolic_variable.h"
 #include "drake/common/symbolic_variables.h"
