+++ drake/solvers/test/symbolic_extraction_test.cc
@@ -9,8 +9,8 @@
 
 #include "drake/common/drake_assert.h"
 #include "drake/common/eigen_matrix_compare.h"
-#include "drake/common/monomial.h"
 #include "drake/common/symbolic_expression.h"
+#include "drake/common/symbolic_monomial.h"
 #include "drake/common/symbolic_polynomial.h"
 #include "drake/common/symbolic_variable.h"
